crypto/cubemedian/requirements.txt
qrcode
numpy
pillow
